feat (calypte_api): merges firmware_info and firmware into one domain
