+++ backend/apps/calypte_api/calypte_api/firmware/api/v1/routers.py
@@ -8,11 +8,10 @@ from calypte_api.common.dependencies import (
 )
 from calypte_api.common.user_roles import UserRole
 from calypte_api.firmware import schemas as firmware_schemas
-from calypte_api.firmware.serivce import FirmwareServiceType
-from calypte_api.firmware_info import schemas as firmware_info_schemas
-from calypte_api.firmware_info.service import FirmwareInfoServiceType
+from calypte_api.firmware.service import FirmwareServiceType
 
 from fastapi import APIRouter, Depends
+from fastapi_pagination import Page
 
 
 router = APIRouter()
@@ -20,7 +19,7 @@ router = APIRouter()
 
 @router.post(
     path="/firmware",
-    response_model=firmware_schemas.UploadFirmwareResponse,
+    response_model=firmware_schemas.CreateFirmwareResponse,
     summary="Upload a firmware",
     description="Upload a firmware",
     response_description="An information of the uploaded firmware",
@@ -29,31 +28,15 @@ router = APIRouter()
 async def upload_firmware(
     _: RateLimiterType,
     create_firmware_request_body: Annotated[
-        firmware_schemas.UploadFirmwareRequestBody, Depends()
+        firmware_schemas.CreateFirmwareRequestBody, Depends()
     ],
-    firmware_info_service: FirmwareInfoServiceType,
     firmware_service: FirmwareServiceType,
     jwt_claims: JwtClaims = Depends(check_permission(UserRole.USER)),
-) -> firmware_schemas.UploadFirmwareResponse:
-    # TODO: combine these two calls into one transaction
-    firmware_info = await firmware_info_service.create_firmware(
+) -> firmware_schemas.CreateFirmwareResponse:
+    return await firmware_service.create_firmware(
         company_id=jwt_claims.user.id,
-        request_body=firmware_info_schemas.CreateFirmwareInfoRequestBody(
-            type_id=create_firmware_request_body.type_id,
-            serial_number=create_firmware_request_body.serial_number,
-            name=create_firmware_request_body.name,
-            version=create_firmware_request_body.version,
-            description=create_firmware_request_body.description,
-        ),
+        request_body=create_firmware_request_body,
     )
-
-    await firmware_service.upload_firmware(
-        company_id=jwt_claims.user.id,
-        firmware_id=firmware_info.id,
-        firmware=create_firmware_request_body.firmware,
-    )
-
-    return firmware_info
 
 
 @router.get(
@@ -70,7 +53,71 @@ async def download_firmware(
     firmware_service: FirmwareServiceType,
     jwt_claims: JwtClaims = Depends(check_permission(UserRole.USER)),
 ) -> firmware_schemas.DownloadFirmwareResponse:
-    return await firmware_service.get_firmware_by_id(
+    return await firmware_service.get_firmware(
         company_id=jwt_claims.user.id,
         firmware_id=firmware_id,
+    )
+
+
+@router.get(
+    path="/firmware/info",
+    response_model=Page[firmware_schemas.GetFirmwareInfoResponse],
+    summary="get firmware info list",
+    description="get a firmware meta data",
+    response_description="the firmware meta data",
+    status_code=200,
+)
+async def retrieve_firmware_list(
+    _: RateLimiterType,
+    firmware_info_service: FirmwareServiceType,
+    query_params: firmware_schemas.GetFirmwareInfoQueryParams = Depends(  # noqa B008
+        firmware_schemas.GetFirmwareInfoQueryParams
+    ),
+    jwt_claims: JwtClaims = Depends(check_permission(UserRole.USER)),  # noqa B008
+) -> Page[firmware_schemas.GetFirmwareInfoResponse]:
+    return await firmware_info_service.get_firmware_info_list(
+        company_id=jwt_claims.user.id,
+        query_params=query_params,
+    )
+
+
+@router.get(
+    path="/firmware/{firmware_id:uuid}/info",
+    response_model=firmware_schemas.GetFirmwareInfoResponse,
+    summary="get an information about the firmware",
+    description="get a firmware meta data",
+    response_description="the firmware meta data",
+    status_code=200,
+)
+async def retrieve_firmware_info(
+    _: RateLimiterType,
+    firmware_id: UUID,
+    firmware_info_service: FirmwareServiceType,
+    jwt_claims: JwtClaims = Depends(check_permission(UserRole.USER)),  # noqa B008
+) -> firmware_schemas.GetFirmwareInfoResponse:
+    return await firmware_info_service.get_firmware_info(
+        company_id=jwt_claims.user.id,
+        firmware_id=firmware_id,
+    )
+
+
+@router.put(
+    path="/firmware/{firmware_id:uuid}/info",
+    response_model=firmware_schemas.GetFirmwareInfoResponse,
+    summary="get an information about the firmware",
+    description="get a firmware meta data",
+    response_description="the firmware meta data",
+    status_code=200,
+)
+async def update_firmware_info(
+    _: RateLimiterType,
+    firmware_id: UUID,
+    update_request_body: firmware_schemas.FirmwareInfoUpdateRequestBody,
+    firmware_info_service: FirmwareServiceType,
+    jwt_claims: JwtClaims = Depends(check_permission(UserRole.USER)),  # noqa B008
+) -> firmware_schemas.UpdateFirmwareInfoResponse:
+    return await firmware_info_service.update_firmware_info(
+        company_id=jwt_claims.user.id,
+        firmware_id=firmware_id,
+        request_body=update_request_body,
     )
