+++ backend/apps/calypte_api/calypte_api/firmware/models.py
