+++ backend/apps/calypte_api/calypte_api/firmware/repositories.py
@@ -1,18 +1,54 @@
 from abc import ABC, abstractmethod
-from typing import Annotated
+from collections.abc import Coroutine, Iterable
+from typing import Any
 from uuid import UUID
 
-from calypte_api.common.dependencies import DBSessionType
-from calypte_api.firmware_info.models import FirmwareInfo
-from calypte_api.firmware_info.schemas import (
-    CreateFirmwareInfoResponse,
+from calypte_api.firmware.models import FirmwareInfo
+from calypte_api.firmware.schemas import (
+    CreateFirmwareResponse,
     GetFirmwareInfoResponse,
     UpdateFirmwareInfoResponse,
 )
 
-from fastapi import Depends
+import aiohttp
+
+from fastapi import UploadFile
+from miniopy_async import Minio
 from sqlalchemy import insert, select, update
 from sqlalchemy.ext.asyncio import AsyncSession
+
+
+class IFirmwareS3Repo(ABC):
+    @abstractmethod
+    async def get_firmware_by_id(
+        self,
+        company_id: UUID,
+        firmware_id: UUID,
+    ) -> Iterable[bytes]:
+        """
+        Get firmware by id
+
+        Args:
+            firmware_id (UUID): firmware id
+            company_id (UUID): user id
+
+        """
+
+    @abstractmethod
+    async def upload_firmware(
+        self,
+        company_id: UUID,
+        firmware_id: UUID,
+        firmware: UploadFile,
+    ) -> None:
+        """
+        Upload firmware
+
+        Args:
+            company_id (UUID): user id
+            firmware_id (UUID): firmware id
+            firmware (bytes): firmware
+        """
 
 
 class IFirmwareInfoRepo(ABC):
@@ -86,7 +122,7 @@ class IFirmwareInfoRepo(ABC):
         name: str,
         description: str,
         version: str,
-    ) -> CreateFirmwareInfoResponse:
+    ) -> CreateFirmwareResponse:
         """
         Create firmware
 
@@ -114,6 +150,52 @@ class IFirmwareInfoRepo(ABC):
             type_id (UUID): type id
             company_id (UUID): company id
         """
+
+
+class FirmwareS3Repo(IFirmwareS3Repo):
+    CHUCK_SIZE_BYTES = 1024 * 1024 * 5  # 5MB
+
+    def __init__(self, client: Minio):
+        self.client = client
+
+    async def get_firmware_by_id(
+        self,
+        company_id: UUID,
+        firmware_id: UUID,
+    ) -> Coroutine[Any, Any, Iterable[bytes]]:
+        async with aiohttp.ClientSession() as session:
+            firmware_info = await self.client.stat_object(
+                bucket_name=str(company_id),
+                object_name=str(firmware_id),
+            )
+            chunk_count = (firmware_info.size // self.CHUCK_SIZE_BYTES) + 1
+            for i in range(chunk_count):
+                response = await self.client.get_object(
+                    bucket_name=str(company_id),
+                    object_name=str(firmware_id),
+                    offset=i * self.CHUCK_SIZE_BYTES,
+                    length=self.CHUCK_SIZE_BYTES,
+                    session=session,
+                )
+                yield await response.read()
+            response.release()
+
+    async def upload_firmware(
+        self,
+        company_id: UUID,
+        firmware_id: UUID,
+        firmware: UploadFile,
+    ) -> None:
+        # bucket = await self.client.bucket_exists(str(company_id))
+        # if not bucket:
+        #     await self.client.make_bucket(str(company_id))
+
+        await self.client.put_object(
+            bucket_name=str(company_id),
+            object_name=str(firmware_id),
+            data=firmware.file,
+            length=firmware.size,
+        )
 
 
 class FirmwareInfoRepo(IFirmwareInfoRepo):
@@ -197,7 +279,6 @@ class FirmwareInfoRepo(IFirmwareInfoRepo):
         )
         firmware_info_model = await self.db_session.scalar(update_stmt)
 
-        await self.db_session.commit()
         return UpdateFirmwareInfoResponse.model_validate(firmware_info_model)
 
     async def create_firmware(
@@ -208,7 +289,7 @@ class FirmwareInfoRepo(IFirmwareInfoRepo):
         name: str,
         description: str,
         version: str,
-    ) -> CreateFirmwareInfoResponse:
+    ) -> CreateFirmwareResponse:
         create_stmt = (
             insert(FirmwareInfo)
             .values(
@@ -226,8 +307,7 @@ class FirmwareInfoRepo(IFirmwareInfoRepo):
 
         firmware_info_model = await self.db_session.scalar(create_stmt)
 
-        await self.db_session.commit()
-        return CreateFirmwareInfoResponse.model_validate(firmware_info_model)
+        return CreateFirmwareResponse.model_validate(firmware_info_model)
 
     async def check_firmware_belongs_to(
         self,
@@ -250,10 +330,9 @@ class FirmwareInfoRepo(IFirmwareInfoRepo):
         return bool(firmware_info_model)
 
 
-def get_firmware_info_repo(db_session: DBSessionType) -> IFirmwareInfoRepo:
+def get_firmware_info_repo(db_session: AsyncSession) -> IFirmwareInfoRepo:
     return FirmwareInfoRepo(db_session=db_session)
 
 
-FirmwareInfoRepositoryType = Annotated[
-    IFirmwareInfoRepo, Depends(get_firmware_info_repo)
-]
+def get_firmware_s3_repo(s3_client: Minio) -> IFirmwareS3Repo:
+    return FirmwareS3Repo(s3_client)
