--- backend/apps/calypte_api/calypte_api/firmware_info/__init__.py
@@ -1,6 +1,0 @@
-from calypte_api.firmware_info.models import FirmwareInfo
-
-
-__all__ = [
-    "FirmwareInfo",
-]
